Role CRUD operations completed. User and Group Features codes updated for relational data.

+++ Users.APP/Features/Groups/GroupDeleteHandler.cs
@@ -16,11 +16,21 @@ namespace Users.APP.Features.Groups
         {
         }
 
+        protected override IQueryable<Group> DbSet()
+        {
+            return base.DbSet().Include(g => g.Users);
+        }
+
         public async Task<CommandResponse> Handle(GroupDeleteRequest request, CancellationToken cancellationToken)
         {
             var existingEntity = await DbSet().SingleOrDefaultAsync(groupEntity => groupEntity.Id == request.Id, cancellationToken);
             if (existingEntity is null)
                 return Error("Group not found!");
+
+            // check if any users exist for the group
+            if (existingEntity.Users.Any())
+                return Error("Group can't be deleted because it has relational users!");
+
             await DeleteAsync(existingEntity, cancellationToken);
             return Success("Group deleted successfully.", existingEntity.Id);
         }
